pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "usb-device-bridge-wsl"
dynamic = ["version"]
description = "USB Device Bridge for WSL — Desktop application to easily attach and manage USB devices in Windows Subsystem for Linux with automatic reconnection support"
requires-python = ">=3.10"
dependencies = [
    "flet>=0.24.0",
    "flet-desktop>=0.24.0; sys_platform == 'win32'",
    "pillow>=10.0.0",
    "pystray>=0.19.0",
]

[project.scripts]
usb-device-bridge = "usb_device_bridge.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["usb_device_bridge*"]

[tool.setuptools-scm]
version_file = "usb_device_bridge/_version.py"
tag_regex = "^[vV](?P<version>[0-9]+\\.[0-9]+\\.[0-9]+(?:\\.[0-9]+)?)$"
